Version 1.3.0 - Add Testing - Added React Testing Library - Added Jest - Created simple test for main component - Added extra config in package.json to ensure image and styles imports - Added new test script for CI - Updated travis.yml file
