+++ .travis.yml
@@ -4,3 +4,4 @@ node_js:
 
 script:
   - npm run lint
+  - npm test
