+++ package.json
@@ -6,7 +6,9 @@
   "scripts": {
     "start": "webpack-dev-server --mode=development",
     "build": "webpack --config webpack.prod.js",
-    "lint": "eslint \"src/**\""
+    "lint": "eslint \"src/**\"",
+    "test": "jest --coverage",
+    "test:watch": "jest --watch"
   },
   "keywords": [
     "gsap",
@@ -30,6 +32,8 @@
     "@babel/preset-env": "^7.11.0",
     "@babel/preset-react": "^7.10.4",
     "@babel/runtime": "^7.11.2",
+    "@testing-library/jest-dom": "^5.11.2",
+    "@testing-library/react": "^10.4.8",
     "autoprefixer": "^9.8.6",
     "babel-eslint": "^10.1.0",
     "babel-loader": "^8.1.0",
@@ -40,6 +44,7 @@
     "eslint-plugin-react": "^7.20.5",
     "file-loader": "^6.0.0",
     "html-webpack-plugin": "^4.3.0",
+    "jest": "^26.2.2",
     "mini-css-extract-plugin": "^0.9.0",
     "node-sass": "^4.14.1",
     "postcss-loader": "^3.0.0",
@@ -53,5 +58,11 @@
     "gsap": "^3.4.2",
     "react": "^16.13.1",
     "react-dom": "^16.13.1"
+  },
+  "jest": {
+    "moduleNameMapper": {
+      "\\.(jpg|ico|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$": "<rootDir>/mocks/fileMock.js",
+      "\\.(css|less|sass|scss)$": "<rootDir>/mocks/fileMock.js"
+    }
   }
 }
